requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.6
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.4.2
google-generativeai>=0.3.2
groq>=0.4.0
wikipedia-api>=0.6.0
httpx>=0.25.0